Fix high Stone Spike damage numbers wrapping
With sufficient madcap crystals, Stone Spike could be doing 256 or more
damage, which would wrap around when stored in `byte spawnDmg`. As this
struct is not part of any savefiles, we can safely expand it to a
`word` to fix the issue.

+++ source/loonyland2/bullet.h
@@ -91,7 +91,7 @@ typedef struct bulletDef_t
 	byte spawnMethod;	// how to spawn, above
 	byte spawnCount;	// number of them to spawn
 	byte spawnRate;		// delay before spawning again (0=just do it once)
-	byte spawnDmg;		// damage of spawned projectiles
+	short spawnDmg;		// damage of spawned projectiles
 	// particles
 	byte partDelay;
 	byte partType;
